Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int x,area;
-    scanf("%d",&x);
+    float x,area;
+    scanf("%f",&x);
     area = 3.14*(x**2);
     printf("Area: %.2f\n", (float)area);
     return 0;
